Add functions for displaying command tree.
This is quick way to observe all commands at
once. It doesn't works fully, because there is
a placeholder in the expandCommandTree, but
all other parts seems works and somehow tested.

One problem, echoed tree can't be copied into the
buffer. So there is commented liberator.log for
this purpose.