Implement dynamic subtrees in the command tree
For this reason new key '_expand_' was introduced.
It should point to function that should returns subtree
same as in main command tree. This subtree will be
inserted instead of key '_expand_'.

Alse the new subtree can contain another key '_expand_'
and it will be expanded too. (The infinity recursion is
possible).

The function "keys" was rewrited to support _expand_.
Also two new functions "isIn" and "getValue" was added
and replaced standart "key in obj" and obj[key] everywhere
in the plugin engine.

The things beguns a bit complicated, but they are work. :)