Simplify things with new getMenuClicker.
New wrapper getMenuClicker provides more easy
way to wrap WebDevelopers functions.  Now all
you need to know is menuId, which is  easy
discovered in the webdeveloper.xul

All subtree "outline" was rewrited in new way.
Many new functions was added.

Thanks to thesnowdog and maxauthority on #vimperator
for the glues.